+++ cassandra/connection.py
@@ -25,6 +25,7 @@ if 'gevent.monkey' in sys.modules:
 else:
     from six.moves.queue import Queue, Empty  # noqa
 
+import six
 from six.moves import range
 
 from cassandra import ConsistencyLevel, AuthenticationFailed, OperationTimedOut
@@ -33,24 +34,15 @@ from cassandra.protocol import (ReadyMessage, AuthenticateMessage, OptionsMessag
                                 StartupMessage, ErrorMessage, CredentialsMessage,
                                 QueryMessage, ResultMessage, decode_response,
                                 InvalidRequestException, SupportedMessage)
-import six
+from cassandra.util import OrderedDict
 
 
 log = logging.getLogger(__name__)
 
-locally_supported_compressions = {}
-
-try:
-    import snappy
-except ImportError:
-    pass
-else:
-    # work around apparently buggy snappy decompress
-    def decompress(byts):
-        if byts == '\x00':
-            return ''
-        return snappy.decompress(byts)
-    locally_supported_compressions['snappy'] = (snappy.compress, decompress)
+# We use an ordered dictionary and specifically add lz4 before
+# snappy so that lz4 will be preferred. Changing the order of this
+# will change the compression preferences for the driver.
+locally_supported_compressions = OrderedDict()
 
 try:
     import lz4
@@ -71,6 +63,18 @@ else:
         return lz4.decompress(byts[3::-1] + byts[4:])
 
     locally_supported_compressions['lz4'] = (lz4_compress, lz4_decompress)
+
+try:
+    import snappy
+except ImportError:
+    pass
+else:
+    # work around apparently buggy snappy decompress
+    def decompress(byts):
+        if byts == '\x00':
+            return ''
+        return snappy.decompress(byts)
+    locally_supported_compressions['snappy'] = (snappy.compress, decompress)
 
 
 MAX_STREAM_PER_CONNECTION = 127
@@ -380,7 +384,22 @@ class Connection(object):
                           locally_supported_compressions.keys(),
                           remote_supported_compressions)
             else:
-                compression_type = next(iter(overlap))  # choose any
+                compression_type = None
+                if isinstance(self.compression, basestring):
+                    # the user picked a specific compression type ('snappy' or 'lz4')
+                    if self.compression not in remote_supported_compressions:
+                        raise ProtocolError(
+                            "The requested compression type (%s) is not supported by the Cassandra server at %s"
+                            % (self.compression, self.host))
+                    compression_type = self.compression
+                else:
+                    # our locally supported compressions are ordered to prefer
+                    # lz4, if available
+                    for k in locally_supported_compressions.keys():
+                        if k in overlap:
+                            compression_type = k
+                            break
+
                 # set the decompressor here, but set the compressor only after
                 # a successful Ready message
                 self._compressor, self.decompressor = \