+++ CHANGELOG.rst
@@ -43,6 +43,11 @@ now:
 =====
 In Progress
 
+Features
+--------
+* Allow a specific compression type to be requested for communications with
+  Cassandra and prefer lz4 if available
+
 Bug Fixes
 ---------
 * Update token metadata (for TokenAware calculations) when a node is removed
@@ -60,6 +65,7 @@ Other
 -----
 * Don't log at ERROR when a connection is closed during the startup
   communications
+* Mke scales, blist optional dependencies
 
 1.1.1
 =====